Fixes compilation error introduced in [12308] (for newer gcc)
git-svn-id: http://svn.mythtv.org/svn/trunk@12317 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/hdhrrecorder.cpp
@@ -355,7 +355,7 @@ void HDHRRecorder::StartRecording(void)
             AdjustFilters();
         }
 
-        unsigned long data_length;
+        size_t data_length;
         unsigned char *data_buffer =
             hdhomerun_video_recv(_video_socket,
                                  VIDEO_DATA_BUFFER_SIZE_1S / 5,
